Osan 4, tehtävät 4.20-4.23

+++ osa4/blogilista/backend/controllers/blogs.js
@@ -1,37 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const { title, author, url, likes = 0 } = request.body
   if(!title || !url) {
     return response.status(400).json(request.body)
   }
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if(!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-
-  //const user = await User.findById(request.body.userId)
-  const user = await User.findById(decodedToken.id)
-  if(!user) {
-    return response.status(400).json({ error: 'could not find user with give id' })
-  }
+  const user = request.user
 
   const blog = new Blog({
     title: title, 
@@ -46,9 +28,17 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if(blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(400).json({ error: 'only person who created the blog is allowed to delete it' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
